requirements.txt
pandas==2.0.3
PyGithub==2.8.1
pytest==8.4.2
tabulate==0.9.0
vcrpy==7.0.0